refactor:hooksにmainフォルダ作成
