+++ app/main/page.tsx
@@ -11,9 +11,9 @@ import PuppyCatAnimation from '../../components/animations/PuppyCatAnimation';
 import BackgroundImage from "../../components/atoms/BackgroundImage"
 import AnchorTemporaryDrawer from "../../components/organisms/menu"
 
-import usePetInfo from '../../hooks/usePetInfo';
-import usePetAnimation from '../../hooks/usePetAnimation';
-import usePetIntervals from '../../hooks/usePetIntervals';
+import usePetInfo from '../../hooks/main/usePetInfo';
+import usePetAnimation from '../../hooks/main/usePetAnimation';
+import usePetIntervals from '../../hooks/main/usePetIntervals';
 
 import { mainTheme } from '../../styles/theme'
 import ThemeWrapper from '../../styles/ThemeWrapper';
